requirements.txt
aiohttp==1.0.5
async-timeout==1.2.0
beautifulsoup4==4.5.3
bs4==0.0.1
chardet==2.3.0
discord.py==0.16.7
multidict==2.1.4
requests==2.13.0
websockets==3.3
